backend/requirements.txt
# Backend dependencies
flask==2.0.1
werkzeug==2.0.3
flask-cors==3.0.10
mysql-connector-python==8.0.26
pyjwt==2.1.0
pyotp==2.6.0
qrcode==7.3.1
python-dotenv==0.19.0
requests==2.26.0

